Changed the temporary dashboard to be fully working, with the option to add a url, and view the stats of a url being the mean and standard deviation of every parameter. This dashboard is temporary as its function will be moved to a GUI in the future.
The server no longer takes input from the user, it instead takes the url that was checked the longest time ago and gives it to the client.
The communication has also been changed to be with multiple clients using sockets, with 'exit' being able to be written in the console of the server which will cause all clients to exit after finishing their current task and then server will exit.

In the future, will add multiprocessing to clients, as well as multiple browser support.